Added install_packages.png screenshot
install_packages.png is a screenshot of RStudio asking if I want to updated packages to the latest version. I added the screenshot to the instructions section as a message that users might encounter.

I also changed the instructions to install handwriter from GitHub instead of CRAN. The bug fix in handwriter that allows handwriterApp and handwriterRF to accept files that don't use the CSAFE naming convention is not available on CRAN yet.